+++ tests/TestCase/Model/Table/PagesCategoriesTableTest.php
@@ -110,13 +110,9 @@ class PagesCategoriesTableTest extends TableTestCase
      */
     public function testFindMethods(): void
     {
-        $this->skipIf(!$this->isMySql());
         $query = $this->Table->find('active');
-        $sql = $query->sql();
+        $this->assertSqlEndsWith('FROM `pages_categories` `Categories` INNER JOIN `pages` `Pages` ON (`Pages`.`active` = :c0 AND `Pages`.`created` <= :c1 AND `Categories`.`id` = `Pages`.`category_id`)', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertInstanceOf(FrozenTime::class, $query->getValueBinder()->bindings()[':c1']['value']);
-
-        $this->skipIfCakeIsLessThan('4.3');
-        $this->assertStringEndsWith('FROM `pages_categories` `Categories` INNER JOIN `pages` `Pages` ON (`Pages`.`active` = :c0 AND `Pages`.`created` <= :c1 AND `Categories`.`id` = `Pages`.`category_id`)', $sql);
     }
 }