+++ tests/TestCase/Model/Table/PostsTableTest.php
@@ -135,9 +135,8 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
         $query = $this->Table->find('forIndex');
         $this->assertArrayKeysEqual(['Categories', 'Tags', 'Users'], $query->getContain());
 
-        $this->skipIf(!$this->isMySql());
         $this->skipIfCakeIsLessThan('4.3');
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_categories` `Categories` ON `Categories`.`id` = `Posts`.`category_id` INNER JOIN `users` `Users` ON `Users`.`id` = `Posts`.`user_id` ORDER BY `Posts`.`created` DESC', $query->sql());
+        $this->assertSqlEndsWith('FROM `posts` `Posts` INNER JOIN `posts_categories` `Categories` ON `Categories`.`id` = `Posts`.`category_id` INNER JOIN `users` `Users` ON `Users`.`id` = `Posts`.`user_id` ORDER BY `Posts`.`created` DESC', $query->sql());
     }
 
     /**
@@ -193,10 +192,9 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
      */
     public function testQueryFromFilter(): void
     {
-        $this->skipIf(!$this->isMySql());
         $this->skipIfCakeIsLessThan('4.3');
         $query = $this->Table->queryFromFilter($this->Table->find(), ['tag' => 'test']);
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`tag` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`)', $query->sql());
+        $this->assertSqlEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`tag` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`)', $query->sql());
         $this->assertEquals('test', $query->getValueBinder()->bindings()[':c0']['value']);
     }
 
@@ -206,18 +204,12 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
      */
     public function testQueryForRelated(): void
     {
-        $this->skipIf(!$this->isMySql());
-
         $query = $this->Table->queryForRelated(4);
-        $sql = $query->sql();
+        $this->assertSqlEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2 AND (`Posts`.`preview`) IS NOT NULL AND `Posts`.`preview` != :c3)', $query->sql());
         $this->assertEquals(4, $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(true, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertInstanceof(FrozenTime::class, $query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertEquals([], $query->getValueBinder()->bindings()[':c3']['value']);
-
-        $this->skipIfCakeIsLessThan('4.3');
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2 AND (`Posts`.`preview`) IS NOT NULL AND `Posts`.`preview` != :c3)', $sql);
-
     }
 
     /**
@@ -226,8 +218,6 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
      */
     public function testQueryForRelatedWithoutImages(): void
     {
-        $this->skipIf(!$this->isMySql());
-
         $query = $this->Table->queryForRelated(4, false);
         $sql = $query->sql();
         $this->assertEquals(4, $query->getValueBinder()->bindings()[':c0']['value']);
@@ -235,6 +225,6 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
         $this->assertInstanceof(FrozenTime::class, $query->getValueBinder()->bindings()[':c2']['value']);
 
         $this->skipIfCakeIsLessThan('4.3');
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2)', $sql);
+        $this->assertSqlEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2)', $sql);
     }
 }