+++ tests/TestCase/Model/Table/UsersTableTest.php
@@ -142,24 +142,23 @@ class UsersTableTest extends TableTestCase
      */
     public function testFindMethods(): void
     {
-        $this->skipIf(!$this->isMySql());
         $query = $this->Table->find('active');
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE (`active` = :c0 AND `banned` = :c1)', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE (`active` = :c0 AND `banned` = :c1)', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
 
         $query = $this->Table->find('banned');
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE `banned` = :c0', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE `banned` = :c0', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
 
         $query = $this->Table->find('pending');
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE (`active` = :c0 AND `banned` = :c1)', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE (`active` = :c0 AND `banned` = :c1)', $query->sql());
         $this->assertFalse($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
 
         $this->skipIfCakeIsLessThan('4.3');
         $query = $this->Table->find('auth');
-        $this->assertStringEndsWith('FROM `users` `Users` INNER JOIN `users_groups` `Groups` ON `Groups`.`id` = `Users`.`group_id`', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` INNER JOIN `users_groups` `Groups` ON `Groups`.`id` = `Users`.`group_id`', $query->sql());
     }
 
     /**
@@ -169,12 +168,10 @@ class UsersTableTest extends TableTestCase
     public function testGetActiveList(): void
     {
         $query = $this->Table->getActiveList();
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE `active` = :c0 ORDER BY `username` ASC', $query->sql());
         $this->assertNotEmpty($query->toArray());
         $fromCache = Cache::read('active_users_list', $this->Table->getCacheName())->toArray();
         $this->assertEquals($fromCache, $query->toArray());
-
-        $this->skipIf(!$this->isMySql());
-        $this->assertStringContainsString('FROM `users` `Users` WHERE `active` = :c0 ORDER BY `username` ASC', $query->sql());
     }
 
     /**
@@ -183,8 +180,6 @@ class UsersTableTest extends TableTestCase
      */
     public function testQueryFromFilter(): void
     {
-        $this->skipIf(!$this->isMySql());
-
         $data = [
             'username' => 'test',
             'group' => 1,
@@ -192,21 +187,21 @@ class UsersTableTest extends TableTestCase
         ];
 
         $query = $this->Table->queryFromFilter($this->Table->find(), $data);
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `active` = :c2 AND `banned` = :c3)', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `active` = :c2 AND `banned` = :c3)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertTrue($query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c3']['value']);
 
         $query = $this->Table->queryFromFilter($this->Table->find(), ['status' => 'pending'] + $data);
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `active` = :c2 AND `banned` = :c3)', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `active` = :c2 AND `banned` = :c3)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c3']['value']);
 
         $query = $this->Table->queryFromFilter($this->Table->find(), ['status' => 'banned'] + $data);
-        $this->assertStringEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `banned` = :c2)', $query->sql());
+        $this->assertSqlEndsWith('FROM `users` `Users` WHERE (`username` like :c0 AND `group_id` = :c1 AND `banned` = :c2)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertTrue($query->getValueBinder()->bindings()[':c2']['value']);