+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 ## 2.30 branch
 ### 2.30.4-RC1
 * fixed a large number of deprecations introduced by CakePHP 4.3;
-* added `TestCase::skipIfCakeIsLessThan()` method;
-* it successfully skips tests that generate a different sql syntax.
+* updated for `me-tools` 2.20.7, now uses `assertSqlEndsNotWith()` and
+    `assertSqlEndsWith()` methods to properly test sql syntax.
+    `TestCase::isMySql()` method has been removed;
+* added `TestCase::skipIfCakeIsLessThan()` method.
 
 ### 2.30.3-beta
 * removed the files that interacted with the `cakephp-database-backup` plugin.